chore: move files to more logical directories
