+++ application/react/_global/FilterBar/AdvancedFilterDropdown/ConditionalFilterDropdownOption/ConditionalFilterDropdownOption.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react-lite';
 import { ConditionalFilterFunction } from "@ui/types/filters";
-import { Button } from '../../Button/Button';
+import { Button } from '../../../Button/Button';
 
 type Props<T> = {
   uuid: string,
@@ -14,7 +14,7 @@ type Props<T> = {
   filtersStore: any
 }
 
-export const ConditionalQuickFilter = observer(<T extends any>(props: Props<T>) => {
+export const ConditionalFilterDropdownOption = observer(<T extends any>(props: Props<T>) => {
   const { uuid, conditionalFilterFunction, textToDisplay, onEnabled, onDisabled, filtersStore } = props;
 
   const enabledConditionalFilters = filtersStore.getConditionalQuickFilters();
