+++ application/react/_global/FilterBar/FilterBar.module.scss
@@ -1,7 +1,5 @@
 @import '../../_styles/shared.module.scss';
 
-// TODO @Gavin: Remove unused styles from this file once the filter logic is extracted to separate components
-
 .searchWrapper {
   height: 34px;
   min-width: 80px;
